Expiration and Rejection

+++ Domain/Features/Orders/Enums/AddressType.cs
@@ -1,6 +1,6 @@
 namespace Domain.Enums
 {
-    public enum AddressType : int
+    public enum AddressType : ushort
     {
         StartPoint = 1,
         StoppingPoint = 2,
